requirements.txt
appdirs
pyyaml
jinja2
